embed-api/requirements.txt
fastapi
sentence-transformers
uvicorn
psycopg2-binary
python-dotenv
sqlalchemy
